% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breed_genotype.R
\name{breed_genotype}
\alias{breed_genotype}
\title{Calculate the number of breedings needed to guarantee a fixed number of
offsprings with a certain probability}
\usage{
breed_genotype(
  confidence_p,
  effective_fertility_p,
  genotypes_p = c(0.25, 0.5, 0.25),
  genotypes_N = c(0, 0, 10),
  only_Fe_Male = FALSE,
  fe_male_p = 0.5,
  litter_mean = NULL,
  litter_sd = NULL,
  binomial_p = NULL,
  offsprings_n_sample = NULL,
  method = "poisson"
)
}
\description{
At the moment only one model is implemented (the textbook).
}
