% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_needed_breedings.R,
%   R/single_genotype_breeding.R
\name{calculate_needed_breedings}
\alias{calculate_needed_breedings}
\title{Calculate the number of breedings needed to guarantee a fixed number of
offsprings with a certain probability}
\usage{
calculate_needed_breedings(
  confidence_p,
  effective_fertility_p,
  n_needed,
  litter_mean = NULL,
  offsprings_n_sample = NULL,
  litter_sd = 2.5,
  binomial_p = NULL,
  method = "festing"
)

calculate_needed_breedings(
  confidence_p,
  effective_fertility_p,
  n_needed,
  litter_mean = NULL,
  offsprings_n_sample = NULL,
  litter_sd = 2.5,
  binomial_p = NULL,
  method = "festing"
)
}
\arguments{
\item{confidence_p}{desired confidence level (probability to achieve 
n_needed offsprings), a number between 0 and 1, excluding 1}

\item{effective_fertility_p}{how likely is an animal to be fertile, a number 
between 0 and 1, including 1}

\item{n_needed}{desired total number of offsprings}

\item{litter_mean}{average number of offsprings one fertile animal produces, 
needed for method values "festing", "binomial" and "poisson"}

\item{offsprings_n_sample}{a sample of the offspring numbers one animal 
produces, needed for method value "empirical"; 
for instance, c(3,1,5,8) is a the data from four animals: The first 
animal had 3 offsprings, the second had 1 offspring, the third animal had 5 
offsprings and the fourth animal had 8 offsprings; 
preferably, this should be a large sample}

\item{litter_sd}{standard deviation of the litter size, needed for method 
value "festing"}

\item{method}{can be "festing", "poisson", "binomial" or "empirical"}

\item{calculation_type}{}
}
\value{
The function returns the number of breedings
needed to guarantee \code{n_needed} offprings with probability no less than
\code{confidence_p}.

The function returns the number of breedings
needed to guarantee \code{n_needed} offprings with probability no less than
\code{confidence_p}.


}
\description{
At the moment only one model is implemented

At the moment only one model is implemented (the textbook).
}
\examples{
calculate_needed_breedings(
    confidence_p=0.95,
    effective_fertility_p=0.6,
    n_needed = 30,
    litter_mean = 7,
    litter_sd = 2.5,
    n_litters=10)


Minimal number of breedings

Calculates the minimal number of animals to be bred in order
to achieve a given number of offsprings with a certain confidence, 
under the one of the models described below.

calculate_needed_breedings(
    confidence_p=0.95,
    effective_fertility_p=0.6,
    n_needed = 30,
    litter_mean = 7,
    litter_sd = 2.5,
    n_litters=10)


Minimal number of breedings

Calculates the minimal number of animals to be bred in order
to achieve a given number of offsprings with a certain confidence, 
under the one of the models described below.

}
