% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_needed_breedings_poisson.R
\name{calculate_needed_breedings_poisson}
\alias{calculate_needed_breedings_poisson}
\title{Minimal number of breedings: Poisson}
\usage{
calculate_needed_breedings_poisson(
  confidence_p,
  effective_fertility_p,
  n_needed,
  litter_mean,
  search_strarting_point = 1
)
}
\arguments{
\item{confidence_p}{desired confidence level (probability to achieve 
n_needed offsprings), a number between 0 and 1, excluding 1}

\item{effective_fertility_p}{how likely is an animal to be fertile, a number between 0 and 1, including 1}

\item{n_needed}{desired total number of offsprings}

\item{litter_mean}{average number of offsprings one fertile animal produces}
}
\value{
returns one value: The minimal number of breedings
}
\description{
Calculates the minimal number of animals to be bred in order
to achieve a given number of offsprings with a certain confidence 
under the following assumptions:
Poisson-distributed number of offsprings per one animal,
Binomially distributed number of fertile animals
}
