% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_needed_breedings_textbook.R
\name{calculate_needed_breedings_textbook}
\alias{calculate_needed_breedings_textbook}
\title{Calculate the number of breedings needed to obtain a fixed numberf of litters with a certain confidence}
\usage{
calculate_needed_breedings_textbook(
  confidence_p,
  effective_fertility_p,
  n_litters,
  textbook_error = FALSE
)
}
\arguments{
\item{confidence_p}{a number between 0 and 1}

\item{effective_fertility_p}{a number between 0 and 1}

\item{n_litters}{an integer >= 1, the number of litters
determined by the function calculate_needed_litters with the
method='textbook' option}
}
\value{
The function returns the number of breedings
needed to achieve  \code{n_litters} with probability no less than \code{confidence_p}.
}
\description{
Here the number of fertile mice \code{n_litters} follows the binomial
distribution with probability of success being \code{effective_fertility_p}
and the number of trials being the number of breedings. The function returns
the number of breedings needed to achieve  \code{n_litters} with probability
no less than \code{confidence_p}.
}
\examples{
calculate_needed_breedings_textbook(
    confidence_p=0.95,
    effective_fertility_p=0.6,
    n_litters=10)

}
