% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_needed_litters_textbook.R
\name{calculate_needed_litters_textbook}
\alias{calculate_needed_litters_textbook}
\title{Calculate the number of litters needed to obtain a fixed numberf of mice
with a certain confidence}
\usage{
calculate_needed_litters_textbook(
  confidence_p,
  litter_mean,
  litter_sd = 2.5,
  n_needed
)
}
\arguments{
\item{confidence_p}{a number between 0 and 1}

\item{litter_mean}{the average litter size of the bred mice}

\item{litter_sd}{standart deviation of the litter size, equals 2.5
by default - as specified in the textbook}

\item{n_needed}{number of baby mice needed to be born}
}
\value{
The function returns the number of litters
needed to obtain \code{n_needed} offsprigns with probability no less than
\code{confidence_p}.
}
\description{
Here the number of offsprings one animal produces, i.e the litter size,
is assumed to follow a Gaussian distribution with mean \code{litter_mean} and
the standard deviation \code{litter_sd}. The function returns
the number of litters needed to obtain \code{n_needed} offsprings
with probability
no less than \code{confidence_p}.
}
\details{
Note: In the textbook, a litter cannot have size 0.
}
\examples{
calculate_needed_breedings_textbook(
    confidence_p=0.95,
    effective_fertility_p=0.6,
    n_litters=10)

}
