% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectBorn.R
\name{expectBorn}
\alias{expectBorn}
\title{Expected total number of animals born in a breeding setup}
\usage{
expectBorn(breedingObj)
}
\arguments{
\item{breedingObj}{an object created by singleGenotype or multiGenotype function}
}
\value{
a plot object, the distribution of the total number of born animals reflected as a histogram
}
\description{
Expected total number of animals born in a breeding setup
}
\examples{
breObj <- multiGenotype(
  confidence_p = 0.9, 
  birth_days = 3, 
  genotypes_p = c(0.25, 0.5, 0.25), 
  genotypes_N = c(10,0,10), 
  sex_distribution = "unimportant",
  litter_average = 7,
  strain = "C56BL/6J")
expectBorn(breObj)

}
