% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiGenotype.R
\name{multiGenotype}
\alias{multiGenotype}
\title{Optimal number of breedings, multiple genotypes setup}
\usage{
multiGenotype(
  confidence_p = 0.9,
  birth_days = 3,
  genotypes_p = c(0.25, 0.5, 0.25),
  genotypes_N = c(0, 0, 10),
  genotype_names = NULL,
  sex_distribution = c("unimportant", "all one sex", "balanced"),
  strain = "Festing",
  litter_average = NULL,
  effective_fertility = NULL
)
}
\arguments{
\item{confidence_p}{}

\item{birth_days}{takes values 1,2,3 or 4: 
the desired offsprings should be born within `birth_days`; 
alternatibely, a number between 0 and 1 specifying the effective fertility 
of the animal}

\item{genotypes_p}{probabilities of a single offspring to have each of the respective genotypes, should sum to 1}

\item{genotypes_N}{required number of offsprings of the respective genotypes}

\item{genotype_names}{names of the genotypes, should be the same length as genotypes_p and genotypes_N}

\item{sex_distribution}{takes values "unimportant", "all one sex", "balanced"; 
reflects the distribution of sexes among the offsprings: If "balanced", 
equal numbers of male and female offspring should be born for each genotype; 
if "all one sex", all of the required offsprings should be either male or female
if "unimportant", the offsprings of any sex suffice. 
Female and male offpsings are born with the same frequency (0.5).}

\item{strain}{mouse strains, currently available are: 
129/SvJa, 
A/J, 
AKR/J, 
BALB/cJ, 
C3H/HeJ, 
C3H/HeOuJ, 
C56BL/6J, 
C57_BL/10SnJ, 
CBA/CaJ, 
DBA/2J, 
FVB/N 
SJL/J, 
Festing (when the strain is unknown, use "Festing" as a general 
textbook example)}

\item{litter_average}{average number of offsprings in a single litter}
}
\value{
object of the breedingMulti class
}
\description{
Calculates the smallest number of breedings needed to yield a desired number 
of offsprings of multiple genotypes with a given probability of success
}
\examples{
multiGenotype(
  confidence_p = 0.9, 
  birth_days = 3, 
  genotypes_p = c(0.25, 0.5, 0.25), 
  genotypes_N = c(10,0,10), 
  sex_distribution = "unimportant",
  litter_average = 7,
  strain = "C56BL/6J")

}
