% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilitiesPlot.R
\name{probabilitiesPlot}
\alias{probabilitiesPlot}
\title{Expected total number of animals born in a breeding setup}
\usage{
probabilitiesPlot(breedingObj)
}
\arguments{
\item{breedingObj}{object created by singleGenotype or multiGenotype function}
}
\value{
a ggplot with the number of matings on the X axis and the 
corresponding confidence in the successful breeding outcome on the Y axis
}
\description{
Expected total number of animals born in a breeding setup
}
\examples{
breObj <- multiGenotype(
  confidence_p = 0.9, 
  birth_days = 3, 
  genotypes_p = c(0.25, 0.5, 0.25), 
  genotypes_N = c(10,0,10), 
  sex_distribution = "unimportant",
  litter_average = 7,
  strain = "C56BL/6J")
probabilitiesPlot(breObj)

}
